revert init message, add fixtures dir
