+++ stac_mjx/__init__.py
@@ -1,1 +1,1 @@
-"""init file."""
+"""This module exposes all high level APIs for stac-mjx."""
