+++ tests/test_main.py
@@ -1,27 +1,6 @@
-import pytest
-from unittest.mock import Mock
 from omegaconf import OmegaConf, DictConfig
-from stac_mjx import utils
 from stac_mjx import main
-
-
-@pytest.fixture
-def mock_omegaconf_load(monkeypatch):
-    mock = Mock(
-        side_effect=[
-            OmegaConf.create({"model": "config"}),
-            OmegaConf.create({"stac": "config"}),
-        ]
-    )
-    monkeypatch.setattr(OmegaConf, "load", mock)
-    return mock
-
-
-@pytest.fixture
-def mock_init_params(monkeypatch):
-    mock = Mock()
-    monkeypatch.setattr(utils, "init_params", mock)
-    return mock
+from tests.fixtures.configs import mock_omegaconf_load, mock_init_params
 
 
 def test_load_configs(mock_omegaconf_load, mock_init_params):
